First Commit. Script is capable of basic functionality.
Available commands:
ADD-USER
POST-ALBUM
ADD-ALBUM

A Moderator first ADD(s)-USER and the POST-ALBUM Event. The user will then ADD-ALBUM. The bot will post the album according to the schedule in the event each week will be the next available user with an album in the queue.

See Command.txt for command syntax